pyproject.toml
[tool.poetry]
name = "django-сore"
version = "0.1.0"
description = ""
authors = ["Pavel Nikulenok <pashaaaa86@gmail.com>"]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
views = "^0.3"
requests = "^2.32.3"
[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
flake8-bugbear = "^24.10.31"
flake8-docstrings = "^1.7.0"
isort = "^5.13.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
mypy = "^1.13.0"
coverage = "^7.6.4"
pre-commit = "^4.0.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.flake8]
max-line-length = 120
ignore = ["D100", "D101", "D102"]
exclude = [".git", "__pycache__", "migrations", "node_modules", ".venv", "*/migrations/*"]
[tool.isort]
profile = "black"
line_length = 88
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "*.py"
[tool.mypy]
ignore_missing_imports = true
[tool.coverage.run]
omit = ["*/migrations/*", "*/tests/*"]
[tool.coverage.report]
show_missing = true